// src/shared/types/Models/Pagination.ts
interface Link {
    label: string;
    active: boolean;
    url: string | null;
}

export interface Pagination<T = unknown> {
    current_page: number;
    data: T[];
    first_page_url: string;
    from: number;
    last_page: number;
    last_page_url: string;
    links: Link[];
    next_page_url: string;
    path: string;
    per_page: number;
    prev_page_url: string | null;
    to: number;
    total: number;
}